+++ grails-app/controllers/org/amcworld/springcrm/SeqNumberStoreInterceptor.groovy
@@ -21,7 +21,9 @@
 package org.amcworld.springcrm
 
 import grails.artefact.Interceptor
+import grails.core.GrailsClass
 import groovy.transform.CompileStatic
+import org.grails.core.artefact.DomainClassArtefactHandler
 
 
 @CompileStatic
@@ -45,35 +47,21 @@ class SeqNumberStoreInterceptor implements Interceptor {
     //-- Public methods -------------------------
 
     /**
-     * Called after the action has been executed.  The method stores the next
-     * available sequence number to the instance which is stored in the model.
-     *
-     * @return  always {@code true}
-     */
-    boolean after() {
-        Map<String, Object> model = getModel()
-        if (model) {
-            def inst = model["${controllerName}Instance".toString()]
-            if (inst instanceof NumberedDomain) {
-                NumberedDomain nd = (NumberedDomain) inst
-                if (nd.number == 0i) {
-                    nd.number = seqNumberService.nextNumber(controllerName)
-                }
-            }
-        }
-
-        true
-    }
-
-    /**
-     * Called before the action is executed.  The method sets the number
-     * parameter to zero if the {@code autoNumber} flag is submitted.
+     * Called before the action is executed.  The method obtains the next
+     * sequence number if the {@code autoNumber} flag has been submitted or the
+     * submitted number is zero.
      *
      * @return  always {@code true}
      */
     boolean before() {
-        if (params.number != null && params.autoNumber) {
-            params.number = 0i
+        if (!params.int('number') || params.autoNumber) {
+            GrailsClass gc = grailsApplication.getArtefactByLogicalPropertyName(
+                DomainClassArtefactHandler.TYPE, controllerName
+            )
+            Class<?> clazz = gc?.clazz
+            if (clazz != null && NumberedDomain.isAssignableFrom(clazz)) {
+                params.number = seqNumberService.nextNumber(controllerName)
+            }
         }
 
         true